+++ cypress/e2e/modals.cy.ts
@@ -23,6 +23,11 @@ describe("modals", () => {
       then(get.fixture("example-style.json")).shouldEqualToStoredStyle();
     });
 
+    it("upload via drag and drop", () => {
+      when.dropExampleFile();
+      then(get.fixture("example-style.json")).shouldEqualToStoredStyle();
+    });
+
     describe("when click open url", () => {
       beforeEach(() => {
         const styleFileUrl = get.exampleFileUrl();
@@ -283,7 +288,7 @@ describe("modals", () => {
 
 
 
-    it("inlcude API key when change renderer", () => {
+    it("include API key when change renderer", () => {
 
       when.click("modal:settings.close-modal");
       when.click("nav:open");