+++ CHANGELOG.md
@@ -14,15 +14,16 @@
 - Add support for sprite object in setting modal
 - Set the correct map view when opening a new style on an empty map
 - Allow root-relative urls in the stylefile
+- Add ability to change the size of the panels
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
 
-- Reset the persisted sidebar layout before Cypress style visits so sidebar resize tests start from a clean state
+- Improved the local style open modal and local file upload.
 - Fixed the Expression editor (for long expressions) being able to be float under other components further down
 - Fixed an issue when clicking on a popup and then clicking on the map again
-- Fix modal close button possition
-- Fixed an issue with the generation of tranlations
+- Fix modal close button position
+- Fixed an issue with the generation of translations
 - Fix missing spec info when clicking next to a property
 - Fix Firefox open file that stopped working due to react upgrade
 - Fix issue with missing bottom error panel